Add more conceptual documentation for maintainers about how recursion works.
